fix deprecation errors for np.float, np.int, np.bool in Numpy 1.24+, see: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ particles/smc_samplers.py
@@ -386,7 +386,7 @@ def view_2d_array(theta):
     """Returns a view to record array theta which behaves
     like a (N,d) float array.
     """
-    v = theta.view(np.float)
+    v = theta.view(float)
     N = theta.shape[0]
     v.shape = (N, -1)
     # raise an error if v cannot be reshaped without creating a copy
